test: keep docs cache proof in the publisher integration lane
